Fix int to size_t usage
We have A LOT of size_t handled as int, this one particularly stood
out because it's compared with a size_t later on but if anyone feels
like working on using size_t properly there's a lot more...

+++ src/include/common_utils.h
@@ -525,7 +525,7 @@ static inline void now(struct timespec *ts)
 
 static inline int strmaxcpy(char *dest, const char *src, size_t dest_size)
 {
-	int len = strlen(src);
+	size_t len = strlen(src);
 
 	if (len >= dest_size)
 		return -1;
